Some random animation

+++ Byte2Eat/OrderHistoryViewController.m
@@ -394,7 +394,7 @@
     cell.testScrollView.showsVerticalScrollIndicator = NO;
     cell.testScrollView.contentSize = CGSizeMake(CGRectGetWidth(cell.frame), CGRectGetHeight(cell.frame));
     cell.labelItemName.text = order.itemName;
-    cell.labelOrderDate.text = [NSString stringWithFormat:@"%d",indexPath.row];
+    cell.labelOrderDate.text = [NSString stringWithFormat:@"%ld",(long)indexPath.row];
     cell.testScrollView.backgroundColor = rowColor[indexPath.row];
 }
 
@@ -453,7 +453,7 @@
     [quantity addAttribute:NSForegroundColorAttributeName value:green range:NSMakeRange(0, quantity.length)];
     [quantity addAttribute:NSFontAttributeName value:[UIFont boldSystemFontOfSize:15] range:NSMakeRange(0, quantity.length)];
 
-    NSMutableAttributedString *cost = [[NSMutableAttributedString alloc] initWithString:[NSString stringWithFormat:@"Rs %i/-", [order.price integerValue] * [order.quantity integerValue]]];
+    NSMutableAttributedString *cost = [[NSMutableAttributedString alloc] initWithString:[NSString stringWithFormat:@"Rs %li/-", [order.price integerValue] * [order.quantity integerValue]]];
     [cost addAttribute:NSShadowAttributeName value:greenShadow range:NSMakeRange(0, cost.length)];
     [cost addAttribute:NSForegroundColorAttributeName value:green range:NSMakeRange(0, cost.length)];
     [cost addAttribute:NSFontAttributeName value:[UIFont boldSystemFontOfSize:15] range:NSMakeRange(0, cost.length)];
